requirements.txt
mlxtend==0.23.1
pandas==2.2.2
pandas_gbq==0.22.0
protobuf==5.27.0
python-dotenv==1.0.1
